# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для веб-панели и внутреннего общения bot↔backend",
        "title": "Anonymus Bot Backend API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "description": "Первый шаг аутентификации — проверка логина/пароля",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Логин оператора",
                "parameters": [
                    {
                        "description": "Логин и пароль",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.loginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.loginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/totp/verify": {
            "post": {
                "description": "Второй шаг аутентификации — выдаёт session_token при успехе",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Подтверждение TOTP-кода",
                "parameters": [
                    {
                        "description": "Pending token и код",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.verifyTOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.verifyTOTPResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.loginRequest": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "api.loginResponse": {
            "type": "object",
            "properties": {
                "pending_token": {
                    "type": "string"
                },
                "session_token": {
                    "type": "string"
                },
                "status": {
                    "description": "\"totp_required\" | \"totp_setup_required\" | \"ok\"",
                    "type": "string"
                },
                "totp_qr_url": {
                    "type": "string"
                },
                "totp_secret": {
                    "type": "string"
                }
            }
        },
        "api.verifyTOTPRequest": {
            "type": "object",
            "required": [
                "code",
                "pending_token"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "pending_token": {
                    "type": "string"
                }
            }
        },
        "api.verifyTOTPResponse": {
            "type": "object",
            "properties": {
                "operator_id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "session_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer \u003cваш_session_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}